+++ src/ts/beatmap/chroma.ts
@@ -1,7 +1,7 @@
 import { Easings } from './easings';
 type LookupMethod = 'Regex' | 'Exact' | 'Contains';
 type Array3DPoint = [number, number, number];
-type ArrayColor = [number, number, number, number?];
+type ArrayColor = [number, number, number, number] | [number, number, number];
 
 export enum ChromaDataEnvAbbr {
     _id = 'Ct',