+++ src/ts/tools/events/unlitBomb.ts
@@ -44,14 +44,19 @@ function inputCheckHandler(this: HTMLInputElement) {
 const unlitBomb = (
     notes: beatmap.note.Note[],
     events: beatmap.event.Event[],
-    bpm: beatmap.bpm.BeatPerMinute
+    bpm: beatmap.bpm.BeatPerMinute,
+    environment: beatmap.environment.EnvironmentName
 ) => {
     if (!events.length) {
         return [];
     }
     const arr: beatmap.note.Note[] = [];
     const eventsLight = events
-        .filter((ev) => beatmap.event.isLightEvent(ev))
+        .filter(
+            (ev) =>
+                beatmap.event.isLightEvent(ev) &&
+                beatmap.environment.EnvironmentEventList[environment].includes(ev._type)
+        )
         .sort((a, b) => a._type - b._type);
     const eventState: {
         [key: number]: {
@@ -120,7 +125,12 @@ const unlitBomb = (
             eventLitTime[ev._type].push([ev._time + fadeTime, false]);
         }
         if (
-            ((ev?._floatValue ?? 1) < 0.25 || beatmap.event.isOff(ev)) &&
+            ((ev?._floatValue ?? 1) < 0.25 ||
+                beatmap.event.isOff(ev) ||
+                (ev._customData?._color &&
+                    ((typeof ev._customData._color[3] === 'number' &&
+                        ev._customData._color[3] < 0.25) ||
+                        Math.max(...ev._customData._color) < 0.25))) &&
             eventState[ev._type].state !== 'off'
         ) {
             eventState[ev._type] = {
@@ -174,7 +184,8 @@ function run(mapSettings: BeatmapSettings, mapSet?: beatmap.map.BeatmapSetData):
     const result = unlitBomb(
         mapSet._data._notes,
         mapSet._data._events,
-        mapSettings._bpm
+        mapSettings._bpm,
+        mapSet._environment
     );
 
     if (result.length) {